feat: Adicionando Funcionalidade via container.

+++ Makefile
@@ -1,10 +1,9 @@
+include ./conf/.env
 
 SHELL 				:= /bin/bash
 DIR_TERRAFORM := ./infra
 DIR_ANSIBLE		:=./ansible
-include ./conf/.env
-
-
+ 
 
 .PHONY: help
 help: 
@@ -14,37 +13,61 @@ help:
 			@echo -e "  make k8s_install\tInstalling Kubernetes/Rahcner with Ansible"
 			@echo -e "  make help\t\tShow this help message\n"
 
-.PHONY: infra_terraform
-infra_terraform:
-			@echo -e "\nCreating infrastructure with Terraform\n"
-			@cd $(DIR_TERRAFORM) && terraform init && terraform validate && terraform apply -auto-approve \
-			
-
-.PHONY: infra_terraform_destroy
-infra_terraform_destroy:
-			@echo -e "\nDestroying infrastructure with Terraform\n"
-			@cd $(DIR_TERRAFORM) && terraform destroy -auto-approve && cd - >/dev/null 2>&1
 
 .PHONY: infra_validate
 infra_validate:
+ifeq ($(MAKETYPE),linux)
 			@echo -e "\nValidating infrastructure with Terraform\n"
 			@cd $(DIR_TERRAFORM) && terraform init && terraform validate
+else ifeq ($(MAKETYPE),docker)
+			@echo -e "\nCreating infrastructure with Docker\n"
+			@docker run --rm -v $(PWD):/app -v -v $HOME/.ssh/id_rsa:/home/deploy/.ssh/id_rsa ceievfa/projetorancher:latest infra_validate
+endif
 
 .PHONY: k8s_install
 k8s_install:
+ifeq ($(MAKETYPE),linux)
 			@echo -e "\nInstalling Kubernetes/Rahcner with Ansible\n"
 			@echo -e "This will install Rancher on the target hosts defined in the Ansible inventory file\n"
 			@cd $(DIR_ANSIBLE) && ansible-playbook -i inventory/hosts install-rancher.yaml -e "target_hosts=rancher" ; cd - >/dev/null 2>&1
+else ifeq ($(MAKETYPE),docker)
+			@echo -e "\Installing Kubernetes/Rahcner with Ansible and Docker\n"
+			@docker run --rm -v $(PWD):/app -v -v $HOME/.ssh/id_rsa:/home/deploy/.ssh/id_rsa ceievfa/projetorancher:latest k8s_install
+endif
+
 
 .PHONY: clean
 clean:
+ifeq ($(MAKETYPE),linux)
 			@echo -e "\nCleaning up the environment\n"
 			@rm -rf $(DIR_TERRAFORM)/.terraform >/dev/null 2>&1 \
 			rm -f $(DIR_TERRAFORM)/terraform.tfstate* >/dev/null 2>&1 \
 			rm -f $(DIR_TERRAFORM)/.terraform.lock* >/dev/null 2>&1
+else ifeq ($(MAKETYPE),docker)
+			@echo -e "\cleanup infrastructure with Docker\n"
+			@docker run --rm -v $(PWD):/app -v -v $HOME/.ssh/id_rsa:/home/deploy/.ssh/id_rsa ceievfa/projetorancher:latest clean
+endif
 
 .PHONY: infra_create
-infra_create: infra_terraform k8s_install
+infra_create:
+ifeq ($(MAKETYPE),linux)
+			@echo -e "\nCreating infrastructure with Terraform\n"
+			@cd $(DIR_TERRAFORM) && terraform init && terraform validate && terraform apply -auto-approve
+
+			@echo -e "\nInstalling Kubernetes/Rahcner with Ansible\n"
+			@echo -e "This will install Rancher on the target hosts defined in the Ansible inventory file\n"
+			@cd $(DIR_ANSIBLE) && ansible-playbook -i inventory/hosts install-rancher.yaml -e "target_hosts=rancher" ; cd - >/dev/null 2>&1
+else ifeq ($(MAKETYPE),docker)
+			@echo -e "\nCreating infrastructure with Docker\n"
+			@docker run --rm -v $(PWD):/app -v -v $HOME/.ssh/id_rsa:/home/deploy/.ssh/id_rsa ceievfa/projetorancher:latest infra_create
+endif
 
 .PHONY: infra_destroy
-infra_destroy: infra_terraform_destroy clean
+infra_destroy:
+ifeq ($(MAKETYPE),linux)
+			@echo -e "\nDestroying infrastructure with Terraform\n"
+			@cd $(DIR_TERRAFORM) && terraform destroy -auto-approve && cd - >/dev/null 2>&1
+else ifeq ($(MAKETYPE),docker)
+			@echo -e "\nDestroy infrastructure with Docker\n"
+			@docker run --rm -v $(PWD):/app -v -v $HOME/.ssh/id_rsa:/home/deploy/.ssh/id_rsa ceievfa/projetorancher:latest infra_destroy
+endif
